Abstract Model Class Game(). Everything else to Multiplayer
