+++ pytimeline/pytimeline/settings/base.py
@@ -34,6 +34,7 @@ INSTALLED_APPS = [
     'engine',
     'django_extensions',
     'singleplayer',
+    'multiplayer',
     'pages',
 ]
