+++ pytimeline/singleplayer/models.py
@@ -1,3 +1,14 @@
 from django.db import models
+from engine.models import Card, Game
 
-# Create your models here.
+DEFAULT_DECK_SIZE = 5
+DEFAULT_HAND_SIZE = DEFAULT_DECK_SIZE
+
+class SinglePlayerGame(Game):
+    n_players = 1
+
+    def start(self):
+        """Inicialize a new single player game."""
+        self.initialize_deck()
+        self.initialize_timeline()
+        self.save()
