+++ pytimeline/templates/multiplayer/card_basic.html
