+++ pytimeline/templates/multiplayer/game_detail.html
