+++ pytimeline/templates/multiplayer/user_game_details.html
