+++ pytimeline/templates/singleplayer/singleplayer.html
@@ -18,5 +18,8 @@
             <a href="start_game"> ¡Empezar! </a>
         </div>
 
+        <h2>Tu mano</h2>
+        <div> hola mundo! {{game.cards_selected}} </div>
+
     </div>
 {% endblock %}
